backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis==5.2.1

# Settings
pydantic-settings==2.7.1

# Modbus (Phase 2)
pymodbus==3.7.4

# HTTP client (Phase 4 — Bitrix)
httpx==0.28.1

# AI Agent (Phase 5 — maintenance manual parsing)
openai>=1.60.0
pypdf>=5.0.0
python-docx>=1.1.0

# File upload (multipart form data)
python-multipart>=0.0.9
